Fix: Revert AccessType, Priority, Status to short to match SMALLINT database columns

+++ backend/TelematicsDataConsole.Core/DTOs/ImeiRestriction/ImeiRestrictionDtos.cs
@@ -11,14 +11,14 @@ public class ImeiRestrictionDto
     public string? DeviceImei { get; set; }
     public int? TagId { get; set; }
     public string? TagName { get; set; }
-    public int? AccessType { get; set; }
-    public int? Priority { get; set; }
+    public short? AccessType { get; set; }
+    public short? Priority { get; set; }
     public string? Reason { get; set; }
     public bool? IsPermanent { get; set; }
     public DateTime? ValidFrom { get; set; }
     public DateTime? ValidUntil { get; set; }
     public string? Notes { get; set; }
-    public int Status { get; set; }
+    public short Status { get; set; }
     public DateTime? CreatedAt { get; set; }
 }
 
@@ -32,9 +32,9 @@ public class CreateImeiRestrictionDto
     public int? TagId { get; set; }
 
     [Required]
-    public int AccessType { get; set; }
+    public short AccessType { get; set; }
 
-    public int? Priority { get; set; }
+    public short? Priority { get; set; }
 
     [MaxLength(255)]
     public string? Reason { get; set; }
@@ -53,8 +53,8 @@ public class UpdateImeiRestrictionDto
 {
     public int? DeviceId { get; set; }
     public int? TagId { get; set; }
-    public int? AccessType { get; set; }
-    public int? Priority { get; set; }
+    public short? AccessType { get; set; }
+    public short? Priority { get; set; }
 
     [MaxLength(255)]
     public string? Reason { get; set; }
@@ -66,6 +66,6 @@ public class UpdateImeiRestrictionDto
     [MaxLength(1024)]
     public string? Notes { get; set; }
 
-    public int? Status { get; set; }
+    public short? Status { get; set; }
 }
